requirements.txt
# --- Orchestration and LLM Routing ---
litellm>=1.40.0           # Universal async LLM client — abstracts ZhipuAI, Groq, Moonshot, Together, NVIDIA NIM
langgraph>=0.2.0          # DAG state machine with Send API, persistent checkpointing, and HITL pause mechanics
langchain-core>=0.3.0     # Required base layer that LangGraph builds on top of

# --- Data Validation ---
pydantic>=2.7.0           # Strict-mode schema enforcement on all LLM outputs; validation errors feed back as retry prompts

# --- Terminal Interface ---
typer>=0.12.0             # CLI commands, subcommands, and argument parsing
rich>=13.9.0              # Live panels, execution trees, syntax highlighting, diff viewer, and progress bars

# --- Semantic Vector Memory ---
sqlite-vec>=0.1.6         # Zero-dependency C SQLite extension for storing and querying vector embeddings locally

# --- Container Management ---
docker>=7.1.0             # Python SDK for the Docker Engine API — used for init-docker, exec_run, and the DAC jail

# --- HTTP and Async Transport ---
httpx>=0.27.0             # Async HTTP used internally by litellm
aiohttp>=3.9.0            # Async HTTP sessions used by the MCP stdio proxy layer

# --- Resilience ---
tenacity>=8.3.0           # Exponential backoff specifically targeting HTTP 429 and 502 errors on every LLM call

# --- Configuration ---
pyyaml>=6.0.1             # Parses the cascading agents.yaml, pipeline.yaml, and cli_config.yaml files

# --- Web Search Tool for the Architect ---
firecrawl-py>=1.1.0       # Firecrawl API client; gives the Architect live documentation retrieval

# --- Utilities ---
python-dotenv>=1.0.0      # Loads .env files for local development outside the container